+++ src/datastore/common.ts
@@ -1113,6 +1113,7 @@ export interface DbPoxCycleSignerStacker {
 interface ReOrgEntities {
   blockHeaders: { index_block_hash: string; block_height: number }[];
   blocks: number;
+  microblockHashes: string[];
   microblocks: number;
   minerRewards: number;
   txs: number;