+++ tests/api/synthetic-stx-txs.test.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { DecodedTxResult, TxPayloadTypeID } from 'stacks-encoding-native-js';
 import { CoreNodeBlockMessage } from '../../src/event-stream/core-node-message';
-import { parseMessageTransaction } from '../../src/event-stream/reader';
+import { CoreNodeMsgBlockData, parseMessageTransaction } from '../../src/event-stream/reader';
 import { parseNewBlockMessage } from '../../src/event-stream/event-server';
 
 // Test processing of the psuedo-Stacks transactions, i.e. the ones that
@@ -20,7 +20,12 @@ describe('synthetic stx txs', () => {
     if (!txMsg) {
       throw new Error(`Cound not find tx ${txid}`);
     }
-    const parsed = parseMessageTransaction(ChainID.Mainnet, txMsg, blockMsg, blockMsg.events);
+    const parsed = parseMessageTransaction(
+      ChainID.Mainnet,
+      txMsg,
+      blockMsg as unknown as CoreNodeMsgBlockData,
+      blockMsg.events
+    );
     if (!parsed) {
       throw new Error(`Failed to parse ${txid}`);
     }
@@ -75,7 +80,12 @@ describe('synthetic stx txs', () => {
     if (!txMsg) {
       throw new Error(`Cound not find tx ${txid}`);
     }
-    const parsed = parseMessageTransaction(ChainID.Mainnet, txMsg, blockMsg, blockMsg.events);
+    const parsed = parseMessageTransaction(
+      ChainID.Mainnet,
+      txMsg,
+      blockMsg as unknown as CoreNodeMsgBlockData,
+      blockMsg.events
+    );
     if (!parsed) {
       throw new Error(`Failed to parse ${txid}`);
     }
@@ -130,7 +140,12 @@ describe('synthetic stx txs', () => {
     if (!txMsg) {
       throw new Error(`Cound not find tx ${txid}`);
     }
-    const parsed = parseMessageTransaction(ChainID.Mainnet, txMsg, blockMsg, blockMsg.events);
+    const parsed = parseMessageTransaction(
+      ChainID.Mainnet,
+      txMsg,
+      blockMsg as unknown as CoreNodeMsgBlockData,
+      blockMsg.events
+    );
     if (!parsed) {
       throw new Error(`Failed to parse ${txid}`);
     }
@@ -234,7 +249,12 @@ describe('synthetic stx txs', () => {
     if (!txMsg) {
       throw new Error(`Cound not find tx ${txid}`);
     }
-    const parsed = parseMessageTransaction(ChainID.Mainnet, txMsg, blockMsg, blockMsg.events);
+    const parsed = parseMessageTransaction(
+      ChainID.Mainnet,
+      txMsg,
+      blockMsg as unknown as CoreNodeMsgBlockData,
+      blockMsg.events
+    );
     if (!parsed) {
       throw new Error(`Failed to parse ${txid}`);
     }
@@ -333,7 +353,7 @@ describe('synthetic stx txs', () => {
       const parsed = parseMessageTransaction(
         ChainID.Mainnet,
         payload.txMsg,
-        payload.blockMsg,
+        payload.blockMsg as unknown as CoreNodeMsgBlockData,
         payload.blockMsg.events
       );
       let txType: 'contract_call' | 'token_transfer' | null;